+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - python: do not pass NULL to ffi.gc if the context can't be created #3818
 - Add read/write timeouts to IMAP sockets #3820
 - Add connection timeout to IMAP sockets #3828
+- Disable read timeout during IMAP IDLE #3826
 
 
 ## 1.102.0